Change folder name and use script setup
