+++ udata_front/theme/gouvfr/assets/js/components/discussions/Thread/Thread.vue
@@ -80,7 +80,7 @@ import Avatar from "../Avatar/Avatar.vue";
 import Author from "../Author/Author.vue";
 import ReadMore from "../../ReadMore/ReadMore.vue";
 import ThreadReply from "../ThreadReply/ThreadReply.vue";
-import MarkAsNoSpam from "../../spam/MarkAsNoSpam.vue";
+import MarkAsNoSpam from "../../MarkAsNoSpam/MarkAsNoSpam.vue";
 import { read_only_enabled } from "../../../config";
 import { formatDate } from "@etalab/data.gouv.fr-components";
 import { computed, defineComponent, ref } from "vue";
